requirements.txt
colour
pyyaml
Pillow